Maintenance
Divide some part from the MyApp
We can set every cuttingplane and can do more cuts.
I create a sorting method which determine how should we draw the atoms
to get a good transparency.

+++ szakdolgozat/szakdoga/approximator.h
@@ -127,12 +127,12 @@ namespace approx {
 		BodyList cut_drawinfo() const {
 			BodyList b1 = compact_drawinfo(*app->last_cut_result().negative());
 			BodyList b2 = compact_drawinfo(*app->last_cut_result().positive());
-			int n = (int)b1.points.size();
+			int n = b1.points.size();
 			b1.points.insert(b1.points.end(), b2.points.begin(), b2.points.end());
 			for (int x : b2.indicies) {
 				b1.indicies.push_back(x+n);
 			}
-			b1.index_ranges.push_back((int)b1.indicies.size());
+			b1.index_ranges.push_back(b1.indicies.size());
 			return b1;
 		}
