+++ szakdolgozat/szakdoga/body.h
@@ -54,7 +54,7 @@ namespace approx{
 		}
 
 		//a test lapszama
-		int size() const { return (int)inds.size(); }
+		int size() const { return inds.size(); }
 	    //az i. lap indexe a taroloban
 		int indicies(size_t i) const { return inds[i]; }
 		//indexlista lekerdezese konstans hozzaferesre
